fix: resolve safety analysis circular imports
