+++ README.md
@@ -1,4 +1,4 @@
-version: 0.2.172
+version: 0.2.175
 Author: Miguel Marina <karel.capek.robotics@gmail.com> - [LinkedIn](https://www.linkedin.com/in/progman32/)
 # AutoML
